refactor: split code in blueprints
