+++ app/products/__init__.py
@@ -1,0 +1,6 @@
+from flask import Blueprint
+
+products_bp = Blueprint(
+    "products", __name__, template_folder="templates", static_folder="static", url_prefix="/products"
+)
+from . import views
