+++ app/products/templates/products.html
@@ -8,7 +8,7 @@
     <ul>
     {% for product in products %}
         <li> {{product["id"]}} -
-            <a href="{{url_for('detail_post', id=product['id'])}}">
+            <a href="{{url_for('products.detail_post', id=product['id'])}}">
                 {{ product["name"] }}
             </a>
         </li>
