+++ app/resume/templates/resume.html
@@ -30,7 +30,7 @@
                 <h1 class="h2">Stepan Nazar</h1>
                 <div class="btn-toolbar mb-2 mb-md-0">
                     <div class="btn-group me-2">
-                        <a href="{{ url_for('contact') }}" class="btn btn-sm btn-outline-secondary">Contact</a>
+                        <a href="{{ url_for('resume.contact') }}" class="btn btn-sm btn-outline-secondary">Contact</a>
                     </div>
                 </div>
             </header>
