+++ app/users/__init__.py
@@ -1,0 +1,6 @@
+from flask import Blueprint
+
+users_bp = Blueprint(
+    "users", __name__, template_folder="templates", static_folder="static"
+)
+from . import views
