+++ app/users/templates/hi.html
@@ -6,10 +6,10 @@
     <nav>
         <ul>
             <li>
-                <a href="{{ url_for('greetings', name='John', age=30) }}">Greet John</a>
+                <a href="{{ url_for('users.greetings', name='John', age=30) }}">Greet John</a>
             </li>
             <li>
-                <a href="{{ url_for('admin') }}">Admin</a>
+                <a href="{{ url_for('users.admin') }}">Admin</a>
             </li>
         </ul>
     </nav>
